Extract a component for the SearchInput and its dropdown
